+++ x/genutil/gentx_test.go
@@ -273,6 +273,7 @@ func (suite *GenTxTestSuite) TestDeliverGenTxs() {
 
 				msg := banktypes.NewMsgSend(addr1, addr2, sdk.Coins{sdk.NewInt64Coin(sdk.DefaultBondDenom, 1)})
 				tx, err := helpers.GenTx(
+					rand.New(rand.NewSource(time.Now().UnixNano())),
 					suite.encodingConfig.TxConfig,
 					[]sdk.Msg{msg},
 					sdk.Coins{sdk.NewInt64Coin(sdk.DefaultBondDenom, 10)},